--- src/midgard-admin/components/endpoint-detail/endpoint-detail.component.html
@@ -1,77 +1,0 @@
-<div class="endpoint-detail">
-  <div class="endpoint-detail__container">
-    <div *ngIf="!swaggerObj">
-      <p>Connecting to bifrost...</p>
-    </div>
-    <div class="endpoint-detail__entries" *ngIf="paths && !paths[0][2]?.includes('get')">
-      <p>Cannot list entries for this endpoint.</p>
-    </div>
-    <div class="endpoint-detail__entries" *ngIf="paths && paths[0][2]?.includes('get')">
-      <div class="endpoint-detail__entries__heading">
-        <h3>{{endpoint}}</h3>
-        <fj-button
-          (click)="onButtonClicked(view)"
-          [shape]="'rounded'"
-          [size]="'small'"
-          [iconMargin]="'0px 5px 0px 0px'"
-          [margin]="'0px 5px'"
-          [iconPosition]="'before'"
-          [icon]="'plus'"
-        >
-          New Entry
-        </fj-button>
-        <div class="endpoint-detail__search-container">
-          <fj-text-input theme="dark" [customStyles]="{'width': '350px'}" [placeholder]="'Filter...'" (keyup)="setFilterValue($event)"> </fj-text-input>
-        </div>
-      </div>
-      <div class="endpoint-detail__entries__table">
-        <fj-table
-          mgCrud
-          #crud="mgCrud"
-          [rows]= "crud.rows | filterByColumnProperties : filterValue : tableOptions.columns"
-          [viewPort]="'450px'"
-          [endpoint]="crudInputs?.endpoint"
-          [idProp]="'id'"
-          [dataProp]="crudInputs?.dataProp"
-          [columns]="tableOptions.columns"
-          theme="dark"
-        >
-          <fj-table-custom-column name="Custom" index="10" flex="1" prop="actions">
-            <ng-template fj-table-header-template>
-              <p>Actions</p>
-            </ng-template>
-            <ng-template let-row="row" fj-table-cell-template>
-              <fj-native-dropdown
-                [size]="'small'"
-                [options]="dropdownOptions"
-                [showDefaultOnClick]="true"
-                [customStyles]="{'marginTop': '0px', 'borderColor': '#484b4d', 'color': 'white','backgroundColor': '#4e5256','width' : '40px', 'padding': '0 0px 0 10px'}"
-                (optionSelected)="dropdownActionTriggered(row, $event)"
-              >
-              </fj-native-dropdown>
-            </ng-template>
-          </fj-table-custom-column>
-        </fj-table>
-      </div>
-    </div>
-    <div class="endpoint-detail__paths" *ngIf="paths?.length > 0">
-      <div class="endpoint-detail__paths__item__overlay" *ngIf="showDefinitions">
-        <img class="endpoint-detail__paths__item__overlay__close" src="assets/img/close-icon.svg" (click)="showDefinitions = false"/>
-        <pre>{{definitions}}</pre>
-      </div>
-      <h3>Paths</h3>
-      <div *ngFor="let path of paths">
-        <div class="endpoint-detail__paths__endpoint" >
-          <h5 class="endpoint-detail__paths__endpoint__text" title="{{path[0]}}">{{path[0]}}</h5>
-          <input type="text" value="{{path[0]}}" hidden>
-          <img class="endpoint-detail__paths__endpoint__icon" (click)="copyEndpoint(path[0])" src="assets/img/copy-icon.svg"/>
-        </div>
-        <div class="endpoint-detail__paths__container">
-          <div *ngFor="let verb of path[2]" class="endpoint-detail__paths__item endpoint-detail__paths__item--{{verb}}">
-            <div class="endpoint-detail__paths__item__verb endpoint-detail__paths__item__verb--{{verb}}" (click)="showDefinitions = !showDefinitions">{{verb}}</div>
-          </div>
-        </div>
-      </div>
-    </div>
-  </div>
-</div>
