+++ src/midgard-admin/components/endpoint-main/endpoint-main.component.html
@@ -1,0 +1,9 @@
+<div class="endpoint-main">
+  <div class="endpoint-main__container">
+    <div *ngIf="!swaggerObj">
+      <p>Connecting to bifrost...</p>
+    </div>
+    <mg-endpoint-entries class="endpoint-main__entries" [definitions]="definitions" [paths]="paths" ></mg-endpoint-entries>
+    <mg-endpoint-paths class="endpoint-main__paths" [definitions]="definitions" [paths]="paths" ></mg-endpoint-paths>
+  </div>
+</div>
