+++ src/midgard-admin/components/endpoint-main/endpoint-main.component.scss
@@ -1,0 +1,36 @@
+@import "../../../styles";
+
+.endpoint-main {
+  display: flex;
+  position: relative;
+  height: 100%;
+  background-color: $dark-background;
+  padding: 30px 0;
+  min-height: 100vh;
+  -webkit-transition: all 0.1s ease-in-out;
+  -moz-transition: all 0.1s ease-in-out;
+  -o-transition: all 0.1s ease-in-out;
+  -ms-transition: all 0.1s ease-in-out;
+  transition: all 0.1s ease-in-out;
+  &__container {
+    display: flex;
+    position: relative;
+    color: $dark-text-color;
+    margin-top: rem-calc(60);
+    width: 100%;
+    -webkit-transition: all 0.1s ease-in-out;
+    -moz-transition: all 0.1s ease-in-out;
+    -o-transition: all 0.1s ease-in-out;
+    -ms-transition: all 0.1s ease-in-out;
+    transition: all 0.1s ease-in-out;
+  }
+  &__entries {
+    flex: 3;
+  }
+  &__paths {
+    flex: 1;
+    max-width: rem-calc(300);
+    margin-left: rem-calc(30);
+    overflow: auto;
+  }
+}
