+++ src/midgard-admin/components/endpoint-main/endpoint-main.component.spec.ts
@@ -1,6 +1,6 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { EndpointDetailComponent } from './endpoint-detail.component';
+import { EndpointDetailComponent } from './endpoint-main.component';
 
 describe('EndpointDetailComponent', () => {
   let component: EndpointDetailComponent;
