+++ src/midgard-admin/components/endpoint-paths/endpoint-paths.component.html
@@ -1,0 +1,19 @@
+<div class="endpoint-paths" *ngIf="paths?.length > 0">
+  <div class="endpoint-paths__item__overlay" *ngIf="showDefinitions">
+    <img class="endpoint-paths__item__overlay__close" src="assets/img/close-icon.svg" (click)="showDefinitions = false"/>
+    <pre>{{definitionsString}}</pre>
+  </div>
+  <h3>Paths</h3>
+  <div *ngFor="let path of paths">
+    <div class="endpoint-paths__endpoint" >
+      <h5 class="endpoint-paths__endpoint__text" title="{{path[0]}}">{{path[0]}}</h5>
+      <input type="text" value="{{path[0]}}" hidden>
+      <img class="endpoint-paths__endpoint__icon" (click)="copyEndpoint(path[0])" src="assets/img/copy-icon.svg"/>
+    </div>
+    <div class="endpoint-paths__container">
+      <div *ngFor="let verb of path[2]" class="endpoint-paths__item endpoint-paths__item--{{verb}}">
+        <div class="endpoint-paths__item__verb endpoint-paths__item__verb--{{verb}}" (click)="showDefinitions = !showDefinitions">{{verb}}</div>
+      </div>
+    </div>
+  </div>
+</div>
