+++ src/midgard-admin/components/endpoint-paths/endpoint-paths.component.scss
@@ -1,0 +1,114 @@
+@import "../../../styles";
+
+$get-color: #0673db;
+$post-color: #02a483;
+$post-color-opacity: rgba(2, 164, 131, 0.17);
+$put-color: #cc9211;
+$delete-color: #fe5a5a;
+
+.endpoint-paths {
+  overflow: auto;
+  &__endpoint {
+    display: flex;
+    &__text {
+      cursor: default;
+      margin: rem-calc(5 0);
+      width: rem-calc(200);
+      white-space: nowrap;
+      overflow: hidden;
+      text-overflow: ellipsis;
+    }
+    &__icon {
+      margin-left: rem-calc(5);
+      cursor: pointer;
+    }
+  }
+  &__container {
+    display: flex;
+    position: relative;
+    margin: rem-calc(15 0);
+  }
+  &__item {
+    display: flex;
+    border-radius: rem-calc(4);
+    margin-left: rem-calc(5);
+    overflow: hidden;
+    &:first-child {
+      margin-left: rem-calc(0);
+    }
+    &__overlay {
+      position: fixed;
+      left: 60%;
+      width: rem-calc(240);
+      padding: rem-calc(15);
+      background-color: $dark-background-brighter;
+      box-shadow: 0px 1px 4px rgba(0, 0, 0, 0.15);
+      z-index: 3;
+      &__close {
+        position: absolute;
+        right: rem-calc(15);
+        cursor: pointer;
+      }
+      pre {
+        overflow: auto;
+        height: rem-calc(450);
+      }
+    }
+    &__verb {
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      font-size: 12px;
+      font-weight: 600;
+      padding: rem-calc(6);
+      width: rem-calc(40);
+      cursor: pointer;
+      &--get {
+        background-color: $get-color
+      }
+      &--put {
+        background-color: $put-color;
+      }
+      &--patch {
+        background-color: $put-color;
+      }
+      &--post {
+        background-color: $post-color;
+      }
+      &--delete {
+        background-color: $delete-color;
+      }
+    }
+    &__path {
+      font-size: 12px;
+      padding: rem-calc(8);
+      overflow: hidden;
+      &__text {
+        color: $dark-text-color;
+        border: 0;
+        padding: 0;
+        width: rem-calc(120);
+        background: transparent;
+        &:focus {
+          outline:none;
+        }
+      }
+    }
+    &--get {
+      border: 1px solid $get-color;
+    }
+    &--put {
+      border: 1px solid $put-color;
+    }
+    &--patch {
+      border: 1px solid $put-color;
+    }
+    &--post {
+      border: 1px solid $post-color;
+      background-color: $post-color-opacity;
+    }
+    &--delete {
+      border: 1px solid #fe5a5a
+    }
+  }
+}
