+++ src/midgard-admin/midgard-admin.module.ts
@@ -9,12 +9,15 @@ import { MainAdminComponent } from './pages/main-admin/main-admin.component';
 import { TopBarAdminComponent } from './components/top-bar-admin/top-bar-admin.component';
 import { NavBarAdminComponent } from './components/nav-bar-admin/nav-bar-admin.component';
 import { NavBarElemAdminComponent } from './components/nav-bar-admin/navbar-elem-admin/navbar-elem-admin.component';
-import { EndpointDetailComponent } from './components/endpoint-detail/endpoint-detail.component';
+import { EndpointMainComponent } from './components/endpoint-main/endpoint-main.component';
 import { MidgardCrudModule } from '../midgard/modules/crud/crud.module';
 import { FilterByColumnPropertiesPipe } from './pipes/filter-by-column-properties.pipe';
 import { ClickOutsideModule } from 'ng-click-outside';
 import { MidgardOAuthModule } from '../midgard/modules/oauth/oauth.module';
 import { AdminAuthGuard } from './guards/auth.guard';
+import { EndpointEntriesComponent } from './components/endpoint-entries/endpoint-entries.component';
+import { EndpointFormComponent } from './components/endpoint-form/endpoint-form.component';
+import { EndpointPathsComponent } from './components/endpoint-paths/endpoint-paths.component';
 
 @NgModule({
   declarations: [
@@ -24,8 +27,11 @@ import { AdminAuthGuard } from './guards/auth.guard';
     TopBarAdminComponent,
     NavBarAdminComponent,
     NavBarElemAdminComponent,
-    EndpointDetailComponent,
-    FilterByColumnPropertiesPipe
+    EndpointMainComponent,
+    FilterByColumnPropertiesPipe,
+    EndpointEntriesComponent,
+    EndpointFormComponent,
+    EndpointPathsComponent
   ],
   imports: [
     CommonModule,
