+++ src/midgard-admin/pages/main-admin/main-admin.component.html
@@ -2,7 +2,7 @@
   <mg-top-bar-admin (menuToggled)="toggleMenu($event)"></mg-top-bar-admin>
   <div class="admin-container">
     <mg-nav-bar-admin [hideMenu]="hideMenu" [swaggerObj]="swaggerObj"></mg-nav-bar-admin>
-    <mg-endpoint-detail class="detail-wrapper" [endpoint]="currentEndpoint" [swaggerObj]="swaggerObj"></mg-endpoint-detail>
+    <mg-endpoint-main class="detail-wrapper" [endpoint]="currentEndpoint" [swaggerObj]="swaggerObj"></mg-endpoint-main>
   </div>
 </div>
